make browser always open
This change makes the web browser always open, even if the ports couldn't
bind. It achieves this by making the frontend/backend not immediately
crash and making the backend wait for the `openBrowser` `Task` to
complete. This is desireable behavior for natively packaging this app for
platforms such as Linux/MacOS, but also for Windows (in the future?).

The app still closes as expected on failure, and developer mode still
does not open the web browser.